+++ src/api/services/index.ts
@@ -1,2 +1,3 @@
 export { createNameService } from './name';
+export { createHttpServer } from './http';
 export { createUdpClient, createUdpServer } from './udp';