+++ types/environment.d.ts
@@ -6,6 +6,10 @@ declare module 'process' {
         NETWORK_IFACE: string;
         BIND_HOST: string;
         BIND_PORT: string;
+        SSL_ENABLED: string;
+        SSL_CA: string;
+        SSL_CERT: string;
+        SSL_KEY: string;
       }
     }
   }